requirements.txt
<<<<<<< HEAD
asgiref==3.3.0
beautifulsoup4==4.9.3
CacheControl==0.12.10
cachetools==4.2.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.0.6
click==8.0.3
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.3
django-cors-headers==3.7.0
django-crontab==0.7.1
django-environ==0.4.5
django-extensions==2.2.5
django-requestlogs==0.2.4
djangorestframework==3.12.2
et-xmlfile==1.0.1
firebase-admin==5.2.0
google==3.0.0
google-api-core==1.31.0
google-api-python-client==2.36.0
google-auth==1.33.1
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-cloud-vision==2.4.1
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.41.0
httplib2==0.20.2
idna==2.10
importlib-metadata==4.8.2
joblib==1.0.0
lxml==4.6.4
msgpack==1.0.3
nsepy==0.8
numpy==1.14.5
oauthlib==3.1.0
openpyxl==3.0.7
packaging==21.0
petl==1.7.2
Pillow==8.0.1
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.7.2
sqlparse==0.4.1
urllib3==1.25.11
Werkzeug==0.16.0
=======
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
Pillow==8.4.0
pytz==2021.1
sqlparse==0.4.1
>>>>>>> c571916195af49b2171c76e667617a10b583c0a5
